// apps/ipron/src/app/features/dn-profile/types/dnProfile.types.ts
export type DnProfileType = '0' | '1';

/** NAT 옵션 — 0=미지정, 1=미사용, 2=rport, 3=route, 4=rport+route */
export type NatOption = '0' | '1' | '2' | '3' | '4';